Revert "Move new versions rate limits to the application"
