+++ src/models/dependency.rs
@@ -1,3 +1,7 @@
+use diesel::deserialize::{self, FromSql};
+use diesel::pg::Pg;
+use diesel::sql_types::Integer;
+
 use crate::models::{Crate, Version};
 use crate::schema::*;
 
@@ -28,10 +32,23 @@ pub struct ReverseDependency {
     pub name: String,
 }
 
-pg_enum! {
-    pub enum DependencyKind {
-        Normal = 0,
-        Build = 1,
-        Dev = 2,
+#[derive(Copy, Clone, Serialize, Deserialize, Debug, FromSqlRow)]
+#[serde(rename_all = "lowercase")]
+#[repr(u32)]
+pub enum DependencyKind {
+    Normal = 0,
+    Build = 1,
+    Dev = 2,
+    // if you add a kind here, be sure to update `from_row` below.
+}
+
+impl FromSql<Integer, Pg> for DependencyKind {
+    fn from_sql(bytes: Option<&[u8]>) -> deserialize::Result<Self> {
+        match <i32 as FromSql<Integer, Pg>>::from_sql(bytes)? {
+            0 => Ok(DependencyKind::Normal),
+            1 => Ok(DependencyKind::Build),
+            2 => Ok(DependencyKind::Dev),
+            n => Err(format!("unknown kind: {}", n).into()),
+        }
     }
 }
