+++ src/schema.rs
@@ -605,7 +605,7 @@ table! {
     /// Representation of the `publish_limit_buckets` table.
     ///
     /// (Automatically generated by Diesel.)
-    publish_limit_buckets (user_id, action) {
+    publish_limit_buckets (user_id) {
         /// The `user_id` column of the `publish_limit_buckets` table.
         ///
         /// Its SQL type is `Int4`.
@@ -624,12 +624,6 @@ table! {
         ///
         /// (Automatically generated by Diesel.)
         last_refill -> Timestamp,
-        /// The `action` column of the `publish_limit_buckets` table.
-        ///
-        /// Its SQL type is `Int4`.
-        ///
-        /// (Automatically generated by Diesel.)
-        action -> Int4,
     }
 }
 
@@ -640,7 +634,7 @@ table! {
     /// Representation of the `publish_rate_overrides` table.
     ///
     /// (Automatically generated by Diesel.)
-    publish_rate_overrides (user_id, action) {
+    publish_rate_overrides (user_id) {
         /// The `user_id` column of the `publish_rate_overrides` table.
         ///
         /// Its SQL type is `Int4`.
@@ -659,30 +653,6 @@ table! {
         ///
         /// (Automatically generated by Diesel.)
         expires_at -> Nullable<Timestamp>,
-        /// The `action` column of the `publish_rate_overrides` table.
-        ///
-        /// Its SQL type is `Int4`.
-        ///
-        /// (Automatically generated by Diesel.)
-        action -> Int4,
-    }
-}
-
-table! {
-    /// Representation of the `recent_crate_downloads` view.
-    ///
-    /// This data represents the downloads in the last 90 days.
-    /// This view does not contain realtime data.
-    /// It is refreshed by the `update-downloads` script.
-    recent_crate_downloads (crate_id) {
-        /// The `crate_id` column of the `recent_crate_downloads` view.
-        ///
-        /// Its SQL type is `Integer`.
-        crate_id -> Integer,
-        /// The `downloads` column of the `recent_crate_downloads` table.
-        ///
-        /// Its SQL type is `BigInt`.
-        downloads -> BigInt,
     }
 }
 
@@ -706,6 +676,24 @@ table! {
         ///
         /// (Automatically generated by Diesel.)
         rendered_at -> Timestamp,
+    }
+}
+
+table! {
+    /// Representation of the `recent_crate_downloads` view.
+    ///
+    /// This data represents the downloads in the last 90 days.
+    /// This view does not contain realtime data.
+    /// It is refreshed by the `update-downloads` script.
+    recent_crate_downloads (crate_id) {
+        /// The `crate_id` column of the `recent_crate_downloads` view.
+        ///
+        /// Its SQL type is `Integer`.
+        crate_id -> Integer,
+        /// The `downloads` column of the `recent_crate_downloads` table.
+        ///
+        /// Its SQL type is `BigInt`.
+        downloads -> BigInt,
     }
 }
