+++ src/util/errors/json.rs
@@ -1,7 +1,6 @@
 use std::fmt;
 
 use super::{AppError, InternalAppErrorStatic};
-use crate::rate_limiter::LimitedAction;
 use crate::util::{json_response, AppResponse};
 
 use chrono::NaiveDateTime;
@@ -91,7 +90,6 @@ pub(super) struct BadRequest(pub(super) String);
 pub(super) struct ServerError(pub(super) String);
 #[derive(Debug)]
 pub(crate) struct TooManyRequests {
-    pub action: LimitedAction,
     pub retry_after: NaiveDateTime,
 }
 
@@ -139,9 +137,9 @@ impl AppError for TooManyRequests {
         let retry_after = self.retry_after.format(HTTP_DATE_FORMAT);
 
         let detail = format!(
-            "{} Please try again after {} or email \
+            "You have published too many crates in a \
+             short period of time. Please try again after {} or email \
              help@crates.io to have your limit increased.",
-            self.action.error_message(),
             retry_after
         );
         let mut response = json_error(&detail, StatusCode::TOO_MANY_REQUESTS);
